Created shared config for externals
